refactor: move `ParsePageRequest` function to dedicated `pagination` module
