+++ src/model/misc.go
@@ -1,7 +1,5 @@
 package model
 
-import "errors"
-
 // DateRange represents a date range with a start and end date.
 type DateRange struct {
 	StartDate string `json:"startDate" form:"startDate" time_format:"2024-10-01" default:""`
@@ -28,45 +26,6 @@ func (pr *PageRequest) GetPageRequest() *PageRequest {
 // to be accepted by ParsePageRequest function
 type PageRequester interface {
 	GetPageRequest() *PageRequest
-}
-
-func parsePageRequestError(err string) (int, int, error) {
-	return 0, 0, errors.New(err)
-}
-
-var (
-	DefaultPage     = 1
-	DefaultPageSize = 10
-)
-
-func ParsePageRequest(pageRequest PageRequester) (int, int, error) {
-	if pageRequest == nil {
-		return DefaultPage, DefaultPageSize, nil
-	}
-
-	pr := (pageRequest).GetPageRequest()
-	if pr == nil {
-		return DefaultPage, DefaultPageSize, nil
-	}
-
-	page, pageSize := pr.Page, pr.PageSize
-
-	if page == nil {
-		page = &DefaultPage
-	}
-	if pageSize == nil {
-		pageSize = &DefaultPageSize
-	}
-
-	if *page <= 0 {
-		return parsePageRequestError("invalid page number")
-	}
-
-	if *pageSize <= 0 {
-		return parsePageRequestError("invalid page size")
-	}
-
-	return *page, *pageSize, nil
 }
 
 // FactHiringProcessFilter represents a filter for querying FactHiringProcess entities.
