+++ src/service/user.go
@@ -5,6 +5,7 @@ import (
 
 	"api5back/ent"
 	"api5back/src/model"
+	"api5back/src/pagination"
 	"api5back/src/processing"
 )
 
@@ -17,7 +18,7 @@ func GetUsers(
 		DimUser.
 		Query()
 
-	page, pageSize, err := model.ParsePageRequest(pageRequest)
+	page, pageSize, err := pagination.ParsePageRequest(pageRequest)
 	if err != nil {
 		return nil, err
 	}
